+++ calculator/src/main/java/runtime/terror/calculator/MyCalculator.java
@@ -2,7 +2,7 @@ package runtime.terror.calculator;
 
 import java.awt.*;
 import java.awt.event.*;
-import javax.swing.JFrame;
+import javax.swing.*;
 
 public class MyCalculator extends JFrame {
 
@@ -132,508 +132,18 @@ public class MyCalculator extends JFrame {
 
     }
 
-
-//////////////////////////////////
-
-
     static String getFormattedText(double temp) {
 
         String resText = "" + temp;
-
-
         if (resText.lastIndexOf(".0") > 0)
-
             resText = resText.substring(0, resText.length() - 2);
-
         return resText;
-
-
     }
-
-
-////////////////////////////////////////
-
 
     public static void main(String[] args) {
-        new MyCalculator("Calculator - JavaTpoint");
-
-    }
-
-
-}
-
-
-/*******************************************/
-
-
-class MyDigitButton extends Button implements ActionListener {
-    MyCalculator cl;
-    MyDigitButton(int x, int y, int width, int height, String cap, MyCalculator clc) {
-
-
-        super(cap);
-
-
-        setBounds(x, y, width, height);
-
-
-        this.cl = clc;
-
-
-        this.cl.add(this);
-
-
-        addActionListener(this);
-
-
-    }
-
-
-////////////////////////////////////////////////
-
-
-    static boolean isInString(String s, char ch) {
-
-
-        for (int i = 0; i < s.length(); i++) if (s.charAt(i) == ch) return true;
-
-
-        return false;
-
-
-    }
-
-
-/////////////////////////////////////////////////
-
-
-    public void actionPerformed(ActionEvent ev) {
-
-
-        String tempText = ((MyDigitButton) ev.getSource()).getLabel();
-
-
-        if (tempText.equals(".")) {
-
-
-            if (cl.Clear) {
-                cl.displayLabel.setText("0.");
-                cl.Clear = false;
-            } else if (!isInString(cl.displayLabel.getText(), '.'))
-
-
-                cl.displayLabel.setText(cl.displayLabel.getText() + ".");
-
-
-            return;
-
-
-        }
-
-
-        int index = 0;
-
-
-        try {
-
-
-            index = Integer.parseInt(tempText);
-
-
-        } catch (NumberFormatException e) {
-            return;
-        }
-
-
-        if (index == 0 && cl.displayLabel.getText().equals("0")) return;
-
-
-        if (cl.Clear) {
-            cl.displayLabel.setText("" + index);
-            cl.Clear = false;
-        } else
-
-
-            cl.displayLabel.setText(cl.displayLabel.getText() + index);
-
-
-    }//actionPerformed
-
-
-}
-
-
-class MyOperatorButton extends Button implements ActionListener {
-
-
-    MyCalculator cl;
-
-
-    MyOperatorButton(int x, int y, int width, int height, String cap, MyCalculator clc) {
-
-
-        super(cap);
-
-
-        setBounds(x, y, width, height);
-
-
-        this.cl = clc;
-
-
-        this.cl.add(this);
-
-
-        addActionListener(this);
-
-
-    }
-
-
-///////////////////////
-
-
-    public void actionPerformed(ActionEvent ev) {
-
-
-        String opText = ((MyOperatorButton) ev.getSource()).getLabel();
-
-
-        cl.Clear = true;
-
-
-        double temp = Double.parseDouble(cl.displayLabel.getText());
-
-
-        if (opText.equals("1/x")) {
-
-
-            try {
-                double tempd = 1 / (double) temp;
-
-
-                cl.displayLabel.setText(MyCalculator.getFormattedText(tempd));
-            } catch (ArithmeticException excp) {
-                cl.displayLabel.setText("Divide by 0.");
-            }
-
-
-            return;
-
-
-        }
-
-
-        if (opText.equals("sqrt")) {
-
-
-            try {
-                double tempd = Math.sqrt(temp);
-
-
-                cl.displayLabel.setText(MyCalculator.getFormattedText(tempd));
-            } catch (ArithmeticException excp) {
-                cl.displayLabel.setText("Divide by 0.");
-            }
-
-
-            return;
-
-
-        }
-
-
-        if (!opText.equals("=")) {
-
-
-            cl.number = temp;
-
-
-            cl.op = opText.charAt(0);
-
-
-            return;
-
-
-        }
-
-
-// process = button pressed
-
-
-        switch (cl.op) {
-
-
-            case '+':
-
-
-                temp += cl.number;
-                break;
-
-
-            case '-':
-
-
-                temp = cl.number - temp;
-                break;
-
-
-            case '*':
-
-
-                temp *= cl.number;
-                break;
-
-
-            case '%':
-
-
-                try {
-                    temp = cl.number % temp;
-                } catch (ArithmeticException excp) {
-                    cl.displayLabel.setText("Divide by 0.");
-                    return;
-                }
-
-
-                break;
-
-
-            case '/':
-
-
-                try {
-                    temp = cl.number / temp;
-                } catch (ArithmeticException excp) {
-                    cl.displayLabel.setText("Divide by 0.");
-                    return;
-                }
-
-
-                break;
-
-
-        }//switch
-
-
-        cl.displayLabel.setText(MyCalculator.getFormattedText(temp));
-
-
-//cl.number=temp;
-
-
-    }//actionPerformed
-
-
-}//class
-
-
-/****************************************/
-
-
-class MyMemoryButton extends Button implements ActionListener {
-
-
-    MyCalculator cl;
-
-
-/////////////////////////////////
-
-
-    MyMemoryButton(int x, int y, int width, int height, String cap, MyCalculator clc) {
-
-
-        super(cap);
-
-
-        setBounds(x, y, width, height);
-
-
-        this.cl = clc;
-
-
-        this.cl.add(this);
-
-
-        addActionListener(this);
-
-
-    }
-
-
-////////////////////////////////////////////////
-
-
-    public void actionPerformed(ActionEvent ev) {
-
-
-        char memop = ((MyMemoryButton) ev.getSource()).getLabel().charAt(1);
-
-
-        cl.Clear = true;
-
-
-        double temp = Double.parseDouble(cl.displayLabel.getText());
-
-
-        switch (memop) {
-
-
-            case 'C':
-
-
-                cl.memLabel.setText(" ");
-                cl.Value = 0.0;
-                break;
-
-
-            case 'R':
-
-
-                cl.displayLabel.setText(MyCalculator.getFormattedText(cl.Value));
-                break;
-
-
-            case 'S':
-
-
-                cl.Value = 0.0;
-
-
-            case '+':
-
-
-                cl.Value += Double.parseDouble(cl.displayLabel.getText());
-
-
-                if (cl.displayLabel.getText().equals("0") || cl.displayLabel.getText().equals("0.0"))
-
-
-                    cl.memLabel.setText(" ");
-
-
-                else
-
-
-                    cl.memLabel.setText("M");
-
-
-                break;
-
-
-        }//switch
-
-
-    }//actionPerformed
-
-
-}//class
-
-
-/*****************************************/
-
-
-class MySpecialButton extends Button implements ActionListener {
-
-
-    MyCalculator cl;
-
-
-    MySpecialButton(int x, int y, int width, int height, String cap, MyCalculator clc) {
-
-
-        super(cap);
-
-
-        setBounds(x, y, width, height);
-
-
-        this.cl = clc;
-
-
-        this.cl.add(this);
-
-
-        addActionListener(this);
-
-
-    }
-
-
-//////////////////////
-
-
-    static String backSpace(String s) {
-
-
-        String Res = "";
-
-
-        for (int i = 0; i < s.length() - 1; i++) Res += s.charAt(i);
-
-
-        return Res;
-
-
-    }
-
-
-//////////////////////////////////////////////////////////
-
-
-    public void actionPerformed(ActionEvent ev) {
-
-
-        String opText = ((MySpecialButton) ev.getSource()).getLabel();
-
-
-//check for backspace button
-
-
-        if (opText.equals("Backspc")) {
-
-
-            String tempText = backSpace(cl.displayLabel.getText());
-
-
-            if (tempText.equals(""))
-
-
-                cl.displayLabel.setText("0");
-
-
-            else
-
-
-                cl.displayLabel.setText(tempText);
-
-
-            return;
-
-
-        }
-
-
-//check for "C" button i.e. Reset
-
-
-        if (opText.equals("C")) {
-
-
-            cl.number = 0.0;
-            cl.op = ' ';
-            cl.Value = 0.0;
-
-
-            cl.memLabel.setText(" ");
-
-
-        }
-
-        cl.displayLabel.setText("0");
-        cl.Clear = true;
-
-
+        new MyCalculator("404Notfound Calculator");
     }
 
 }
+
+
